Allow parent scroll to scroll when child reaches its end
When a scrollable component is nested inside another scrollable component, the parent must be able to scroll when the child reaches its end. Otherwise it may be impossible to scroll the parent in certain situations (ex. when a child fills the screen entirely). The ArkUI builtin settings do not handle the case well when the child scroll is inverted - therefore this had to be added manually.

closes #691, #685

See merge request rnoh/react-native-harmony!638